fix: move asyncGenerate into ring0, since it needs Symbol
