+++ packages/stdlib/src/ring0/main.js
@@ -1,1 +1,2 @@
 export * from './makers.js';
+export * from './async-generate.js';
